Update show.html.erb
The main issue was the the topic_id was not included in the :board params so rails didnt know what to do with it.  the (board: { topic_id: topic.id }) fixes this by creating a nested params that the controller can recognize as belonging to boards.  the controller accesses this through params[:board][:topic_id]